rk2 fix. WebPacks SW adaptation
