+++ package.json
@@ -7,8 +7,8 @@
     "lint": "npx eslint . && npx stylelint --fix **/*.scss",
     "debug-no-webpack": "node ./dev/toggleGlobals.js true && node ./dev/compileTemplates.js debug && npm run start-static",
     "debug": "node ./dev/toggleGlobals.js true && node ./dev/compileTemplates.js debug && npx webpack --mode development && npm run start-static",
-    "build": "npx eslint . && node ./dev/toggleGlobals.js false && node ./dev/compileTemplates.js && node ./dev/createCachedList && npx webpack && npm run start-static",
-    "nohup": "npx eslint . && node ./dev/toggleGlobals.js false && node ./dev/compileTemplates.js && node ./dev/createCachedList && npx webpack && nohup node server/server.js",
+    "build": "npx eslint . && node ./dev/toggleGlobals.js false && node ./dev/compileTemplates.js && npx webpack && npm run start-static",
+    "nohup": "npx eslint . && node ./dev/toggleGlobals.js false && node ./dev/compileTemplates.js && npx webpack && nohup node server/server.js",
     "start": "concurrently --kill-others \"npm run start-static\" \"npm run start-api\"",
     "start-static": "node server/server.js",
     "start-api": "node server/server-api.js"
@@ -47,6 +47,7 @@
     "url-loader": "^4.1.1",
     "webpack": "^5.64.4",
     "webpack-cli": "^4.9.1",
+    "workbox-webpack-plugin": "^6.4.2",
     "ws": "^8.3.0"
   }
 }
