+++ public/static/styles/pages/_categories.scss
@@ -48,7 +48,7 @@ $categories-margin: 20px;
   }
 
   &__mark {
-    color: $color-secondary;
+    color: $color-elements;
   }
 
   &__choose-bar {
