+++ webpack.config.js
@@ -3,6 +3,7 @@ const HTMLWebpackPlugin = require('html-webpack-plugin');
 const {CleanWebpackPlugin} = require('clean-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const CopyPlugin = require('copy-webpack-plugin');
+const WorkboxPlugin = require('workbox-webpack-plugin');
 
 module.exports = {
   mode: 'production',
@@ -16,6 +17,31 @@ module.exports = {
     assetModuleFilename: 'static/[name][ext][query]',
   },
   plugins: [
+    new WorkboxPlugin.GenerateSW({
+      // https://developers.google.com/web/tools/workbox/guides/generate-service-worker/webpack
+      // Do not precache images
+      // exclude: [/\.(?:png|jpg|jpeg|svg)$/],
+
+      // Define runtime caching rules.
+      // runtimeCaching: [{
+      //   // Match any request that ends with .png, .jpg, .jpeg or .svg.
+      //   urlPattern: /\.(?:png|jpg|jpeg|svg)$/,
+
+      //   // Apply a cache-first strategy.
+      //   handler: 'CacheFirst',
+
+      //   options: {
+      //     // Use a custom cache name.
+      //     cacheName: 'images',
+
+      //     // Only cache 10 images.
+      //     expiration: {
+      //       maxEntries: 10,
+      //     },
+      //   },
+      // }],
+    }),
+
     new CopyPlugin({
       patterns: [
         {
